services/question-service/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg[binary]>=3.2.0

# Data validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Authentication (token validation only)
python-jose[cryptography]>=3.3.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2